+++ web/apps/photos/src/components/Sidebar/index.tsx
@@ -17,7 +17,19 @@ import {
     TRASH_SECTION,
 } from "@/new/photos/services/collection";
 import type { CollectionSummaries } from "@/new/photos/services/collection/ui";
+import {
+    hasAddOnBonus,
+    hasExceededStorageQuota,
+    isSubscriptionActive,
+    isSubscriptionActiveFree,
+    isSubscriptionActivePaid,
+    isSubscriptionCancelled,
+    isSubscriptionPastDue,
+    isSubscriptionStripe,
+    redirectToCustomerPortal,
+} from "@/new/photos/services/plan";
 import { isInternalUser } from "@/new/photos/services/settings";
+import type { UserDetails } from "@/new/photos/services/user";
 import { isFamilyAdmin, isPartOfFamily } from "@/new/photos/services/user";
 import { AppContext, useAppContext } from "@/new/photos/types/context";
 import { initiateEmail, openURL } from "@/new/photos/utils/web";
@@ -65,19 +77,7 @@ import React, {
 import { Trans } from "react-i18next";
 import { getUncategorizedCollection } from "services/collectionService";
 import exportService from "services/export";
-import {
-    hasAddOnBonus,
-    hasExceededStorageQuota,
-    isSubscriptionActive,
-    isSubscriptionActiveFree,
-    isSubscriptionActivePaid,
-    isSubscriptionCancelled,
-    isSubscriptionPastDue,
-    isSubscriptionStripe,
-    redirectToCustomerPortal,
-} from "services/plan";
 import { getUserDetailsV2 } from "services/userService";
-import { UserDetails } from "types/user";
 import { testUpload } from "../../../tests/upload.test";
 import { MemberSubscriptionManage } from "../MemberSubscriptionManage";
 import { Preferences } from "./Preferences";
