+++ web/apps/photos/src/components/pages/gallery/PlanSelector.tsx
@@ -1,8 +1,27 @@
 import { genericRetriableErrorDialogAttributes } from "@/base/components/utils/dialog";
 import log from "@/base/log";
+import type {
+    Plan,
+    PlanPeriod,
+    PlansData,
+    Subscription,
+} from "@/new/photos/services/plan";
+import billingService, {
+    getPlansData,
+    hasAddOnBonus,
+    isSubscriptionActive,
+    isSubscriptionActiveFree,
+    isSubscriptionActivePaid,
+    isSubscriptionCancelled,
+    isSubscriptionStripe,
+    redirectToCustomerPortal,
+    redirectToPaymentsApp,
+} from "@/new/photos/services/plan";
 import {
+    BonusData,
     getTotalFamilyUsage,
     isPartOfFamily,
+    UserDetails,
 } from "@/new/photos/services/user";
 import { AppContext } from "@/new/photos/types/context";
 import { bytesInGB, formattedStorageByteSize } from "@/new/photos/utils/units";
@@ -35,21 +54,8 @@ import Typography from "@mui/material/Typography";
 import { t } from "i18next";
 import { useContext, useEffect, useMemo, useState } from "react";
 import { Trans } from "react-i18next";
-import type { Plan, PlanPeriod, PlansData, Subscription } from "services/plan";
-import billingService, {
-    getPlansData,
-    hasAddOnBonus,
-    isSubscriptionActive,
-    isSubscriptionActiveFree,
-    isSubscriptionActivePaid,
-    isSubscriptionCancelled,
-    isSubscriptionStripe,
-    redirectToCustomerPortal,
-    redirectToPaymentsApp,
-} from "services/plan";
 import { getFamilyPortalRedirectURL } from "services/userService";
 import { SetLoading } from "types/gallery";
-import { BonusData, UserDetails } from "types/user";
 
 interface PlanSelectorProps {
     modalView: boolean;
