+++ web/apps/photos/src/services/userService.ts
@@ -1,13 +1,13 @@
 import { putAttributes } from "@/accounts/api/user";
 import log from "@/base/log";
 import { apiURL, familyAppOrigin } from "@/base/origins";
+import type { UserDetails } from "@/new/photos/services/user";
 import { getLocalFamilyData, isPartOfFamily } from "@/new/photos/services/user";
 import { ApiError } from "@ente/shared/error";
 import HTTPService from "@ente/shared/network/HTTPService";
 import { LS_KEYS, getData } from "@ente/shared/storage/localStorage";
 import { getToken } from "@ente/shared/storage/localStorage/helpers";
 import { HttpStatusCode } from "axios";
-import { UserDetails } from "types/user";
 
 const HAS_SET_KEYS = "hasSetKeys";
